// src/entities/Profile/model/types/profile.ts
import type { Currency } from '@/entities/Currency'
import type { Country } from '@/entities/Country'

export interface Profile {
    id?: string
    first?: string
    lastname?: string
    age?: number
    currency?: Currency
    country?: Country
    city?: string
    username?: string
    avatar?: string
}